Fundoo App

+++ src/main/java/com/bridgelabz/fundoo/user/respository/UserRespository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bridgelabz.fundoo.user.model.User;
 
-public interface UserRespository extends JpaRepository<User,Integer>{
+public interface UserRespository extends JpaRepository<User,Long>{
 	public Optional<User> findByEmailId(String emailId);
 	
 }
